*** empty log message ***

+++ Hum/Ace/DotterLauncher.pm
@@ -62,27 +62,31 @@ sub fork_dotter {
         return 1;
     }
     elsif (defined $pid) {
-        my $prefix = "/tmp/dotter.$$";
-        my $query_file   = "$prefix.query";
-        my $subject_file = "$prefix.subject";
-        my $query_seq = $seq->sub_sequence($start, $end);
-        $query_seq->name($seq->name);
-        
-        # Write out the query sequence
-        my $query_out = Hum::FastaFileIO->new("> $query_file");
-        $query_out->write_sequences($query_seq);
-        $query_out = undef;
-        
-        # Write the subject with pfetch
-        my ($subject_seq) = Hum::Pfetch::get_Sequences($subject_name);
-        my $subject_out = Hum::FastaFileIO->new("> $subject_file");
-        $subject_out->write_sequences($subject_seq);
-        $subject_out = undef;
-        
-        # Run dotter
-        my $offset = $start - 1;
-        my $dotter_command = "dotter -q $offset $query_file $subject_file ; rm $query_file $subject_file";
-        exec($dotter_command) or warn "Failed to exec '$dotter_command' : $!";
+        eval{
+            my $prefix = "/tmp/dotter.$$";
+            my $query_file   = "$prefix.query";
+            my $subject_file = "$prefix.subject";
+            my $query_seq = $seq->sub_sequence($start, $end);
+            $query_seq->name($seq->name);
+
+            # Write out the query sequence
+            my $query_out = Hum::FastaFileIO->new("> $query_file");
+            $query_out->write_sequences($query_seq);
+            $query_out = undef;
+
+            # Write the subject with pfetch
+            my ($subject_seq) = Hum::Pfetch::get_Sequences($subject_name);
+            die "Can't fetch '$subject_name'\n" unless $subject_seq;
+            my $subject_out = Hum::FastaFileIO->new("> $subject_file");
+            $subject_out->write_sequences($subject_seq);
+            $subject_out = undef;
+
+            # Run dotter
+            my $offset = $start - 1;
+            my $dotter_command = "dotter -q $offset $query_file $subject_file ; rm $query_file $subject_file";
+            exec($dotter_command) or warn "Failed to exec '$dotter_command' : $!";
+        };
+        exit(0);
     }
     else {
         confess "Can't fork: $!";
